Adding job processing to the react web pages.
This adds a simple react component that shows known tasks.  When a task is selected, the appropriate input fields are shown.  Currently, there are two hard-coded IDs: one for the folder where source items are located, and one for the folder where data will be written.

I've changed the job spec yml to provide information necessary for the UI.

I've altered how job specs are imported so that they can be enumerated via the GET osumo/tasks endpoint.

I've altered the POST osumo endpoint so that it can process arbitrary tasks.