Added second part of iGPSe task.
The tasks don't automatically chain together, yet.

Made it so we can show a list of items, but send the appropriate file to the job instead.  Otherwise, if the item and file are not named the same, the girder worker plugin downloads the item as a directory, but expects to get a file.  I think that this could (should?) be changed in the girder worker plugin, as for a one-file item, I don't see the desired default being the current behavior.