Select the new value from the file selector.
This isn't necessarily the best way to do it, but it addresses a number of issues:

- by storing too much in state during componentWillMount, the current selection stops coming from the props.

- there was no key on the Select element (and a needless containing div)

- when a new item is added to the list (or, more strictly, when an item is shown in the list), a call is made to set the parent's state to switch to that item.

You could add arbitrary items, rather than just ones that are already in the file list, by creating a whole new item as needed.  In this case, we could check if the selected item is in the options list and, if not, add it at the end.