provide better job results for osumo/results/:id endpoint
Updates the document returned for osumo/results/:id to have a "results"
attribute instead of a "files" attribute.  The new results attribute now
maps output variable names to their computed results, with additional
data that lets clients pick out results that were written to girder from
inline results.  That also means that we don't have to turn *every*
result we want to use into a girder item, anymore!

Note that there's several moving parts needed to make this work on the
back end, including a plugin for girder-worker that's needed to work
around the disaster that is the current implementation of
girder-worker's mongodb IO mode. (Note: someone, probably me, should
really upstream a fix for it).  The aformentioned plugin lives in
github.com/osumo/sumo_io.