add server-side handling of scratch spaces
Adds server-side handling of scratch spaces.  Now,
every job created through OSUMO will have a
dedicated folder created for its outputs.  This
logic was moved server-side so that the directory
could be created after the job entry was created
(so that we know what id to use), but before
populating the job entry with the girder_io
entries for any specified outputs (that would use
the created folder).  Only the server can address
this chicken-and-egg issue.

Explicitly-specified destination folders are
silently ignored for backwards compatibility.
There are more changes that should be made to the
client to take advantage of the new server-side
feature/eliminate the need for the
silently-ignored parent parameters, but I stop
here to keep the code changes minimal; this PR
should represent the near-minimum changes required
to work around the issue mentioned in #56.