fix issue with feature-apply task script
Fixes an issue where the feature-apply task script
would apply the given reindexing map to both the
files' rows *and* columns.  The script should only
ever apply the map to only a file's rows or its
columns, but not both.  Note that in the current
implementation, the script should always apply the
map to the columns of both input files.